build: Break buildscripts up into multiple files
