+++ build-logic/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id "java-gradle-plugin"
+    id "org.spongepowered.gradle.sponge.dev" version "1.0.2"
+}
+
+gradlePlugin {
+    plugins {
+        impl {
+            id = "implementation-structure"
+            implementationClass = "org.spongepowered.gradle.impl.SpongeImplementationPlugin"
+        }
+    }
+}
